Update README.md
Dive in to explore the intricate synergy between insightful documentation and the executable core of our research This repository encompasses the collaborative research efforts of our team on Labin. Inside, you'll discover a comprehensive PDF report, the corresponding working code saved as a .py file, and an additional PDF showcasing the output of our code.